+++ models/decoders/pure_scn.py
@@ -207,7 +207,7 @@ class PureSCN(nn.Module):
                     -1).topk(k, 0, True, True)
 
             # Convert unrolled indices to actual indices of scores
-            prev_word_inds = top_k_words / vocab_size  # (s)
+            prev_word_inds = (top_k_words / vocab_size).long()  # (s)
             next_word_inds = top_k_words % vocab_size  # (s)
 
             # Add new words to sequences
